+++ klibs/KLEventInterface.py
@@ -2,7 +2,6 @@
 __author__ = 'Jonathan Mulle & Austin Hurst'
 
 from klibs.KLConstants import TK_S, TK_MS
-from klibs import P
 from klibs.KLNamedObject import NamedObject
 from klibs.KLTime import precise_time as time
 from klibs.KLUserInterface import ui_request
@@ -11,6 +10,8 @@ from klibs.KLUserInterface import ui_request
 class TrialEventTicket(NamedObject):
 	"""A ticket defining the name and onset time of an event that occurs within a trial. For use
 	with the :obj:`EventManager` class.
+
+	Deprecated, do not use.
 
 	Args:
 		label (:obj:`str`): The identfying name of the trial event (e.g. 'cue on').
@@ -55,212 +56,207 @@ class TrialEventTicket(NamedObject):
 		return self.__onset
 
 
+
 class EventManager(object):
-	"""A class for sequencing the events that will occur during a given trial, relative to the
-	trial onset.
+	"""A class for defining and controlling the sequence of events in a trial.
 
-	.. note:: An EventManager object is provided globally in the KLibs runtime environment as the
-	  experiment attribute ``self.evm``. This should be used instead of creating your own instance.
-
-	The EventManager object works in two stages:
+	During a given trial of a task, there are usually `things that happen`: for
+	example, a cue might appear on-screen after 1000 ms and disappear 100 ms later,
+	or a target might appear after a random delay between 1500 and 3000 ms. The
+	``EventManager`` class exists to make sequencing those sorts of events within
+	a trial simple and easy.
 	
-	First, before the start of a trial, the names and onset times of the events that should occur
-	within the trial are registered with the EventManager using the :meth:`register_ticket` or
-	:meth:`register_tickets` methods, e.g.::
+	First, before the trial begins, we add the names and onsets of the trial's
+	events to the ``EventManager`` using the :meth:`add_event` method, e.g.::
+			
+			evm = EventManager()
+			evm.add_event('cue_on', 1000)
+			evm.add_event('cue_off', 100, after='cue_on')
+			evm.add_event('target_on', 400, after='cue_off')
 
-			# Note: events[-1][1] means the onset time of the previous event in the list
-			events = []
-			events.append(['cue_on', 1000])
-			events.append(['cue_off', events[-1][1] + 100])
-			events.append(['target_on', events[-1][1] + 400])
-			self.evm.register_tickets(events)
+	Then, during the trial itself, we can use the :meth:`before`, :meth:`after`,
+	and :meth:`between` methods to make decisions based on whether those events
+	have occurred::
 
-	Then, in the body of the experiment's :meth:`~.KLExperiment.trial` method, the EventManager can
-	then be used to watch for those events and change behaviour when they occur::
-
-		while self.evm.before('target_on', pump_events=True):
+		while evm.before('target_on'):
+			ui_request()
 			fill()
 			draw_stimuli()
-			if self.evm.between('cue_on', 'cue_off'):
+			if evm.between('cue_on', 'cue_off'):
 				draw_cue()
 			flip()
 
-	This is done to make it easy to group all trial event sequencing information in one place in
-	your code, as well as make for simple and human-readable code within the trial body.
+	The klibs Experiment class provides an ``EventManager`` instance as ``self.evm``,
+	so you generally won't need to create one yourself. Its trial clock is started
+	automatically when ``self.trial()`` is called on each trial, and is reset
+	after every trial ends.
 
 	"""
 	def __init__(self):
 		super(EventManager, self).__init__()
 		self.events = {}
+		self._issued = {}
 		self.start_time = None
 
 
-	def __event_issued(self, label):
-		if self.events[label].issued == True:
-			return True
-		elif self.events[label].onset < self.trial_time_ms:
-			self.events[label].issued = True
-			return True
-		else:
-			return False
+	def _ensure_exists(self, label):
+		# Makes sure a given event exists in the manager, raising an exception
+		# if it doesn't
+		try:
+			self.events[label]
+		except KeyError:
+			err = "'{0}' does not match the name of any existing event."
+			raise ValueError(err.format(label))
+
+
+	def add_event(self, label, onset, after=None):
+		"""Adds an event to the event manager.
+		
+		By default, the onsets of events are relative to the start of the trial
+		(e.g. a 'cue_on' event with an onset of 1000 will occur 1000 ms after the
+		trial begins). However, you can also define an event's onset as being
+		relative to another event's onset using the optional 'after' argument
+		(e.g. a 'cue_off' event occuring 100 ms after the 'cue_on' event).
+		
+		Args:
+			label (str): The name of the event to add (e.g. 'cue_on').
+			onset (int): The onset of the event (in milliseconds), relative to
+				the start of the trial (by default) or another event (if 'after'
+				is specified).
+			after (str, optional): The name of an existing event that the onset
+				is relative to. If not specified, the onset is relative to the
+				start of the trial.
+
+		"""
+		if after is not None:
+			# If 'after' provided, 'onset' is relative to that event
+			self._ensure_exists(after)
+			onset = self.events[after] + onset
+		self.events[label] = onset
+		self._issued[label] = False
 
 	
 	def register_ticket(self, event):
-		"""Registers an event with the EventManager for the upcoming trial.
-
-		If defining an event with a :obj:`List`, it must be in the format ``[label, onset]``, 
-		with 'label' being the name of the event and 'onset' being the time (in millseconds)
-		after trial start that the event occurs, such as::
-
-			self.evm.register_ticket(['target_on', 1000]) # 1000ms after trial start
-
-		Args:
-			event (:obj:`TrialEventTicket` or :obj:`List`): An event or List defining an
-				event to be registered with the EventManager.
-
-		"""
+		# Deprecated, use add_event instead
 		if not isinstance(event, TrialEventTicket):
 			try:
 				event = TrialEventTicket(*event)
 			except SyntaxError:
 				raise TypeError("Argument 'event' must be a List or a TrialEventTicket.")
-		self.events[event.label] = event
+		self.add_event(event.label, event.onset)
 
 
 	def register_tickets(self, events):
-		"""Registers multiple events with the EventManager for the upcoming trial.
-
-		See also: :meth:`register_ticket`
-
-		Args:
-			events (:obj:`List`): A List of :obj:`TrialEventTicket` or :obj:`List` objects
-				defining events to be registered with the EventManager. 
-
-		"""
+		# Deprecated, use add_event instead
 		for e in events:
 			self.register_ticket(e)
 
 
-	def registered(self, label):
-		"""Checks if an event with a given name is currently registered with the EventManager.
-
-		Args:
-			label (str): The label of the event to check the registration of.
-
-		Returns:
-			True if the event is registered with the EventManager, otherwise False.
-
-		"""
-		return label in self.events
-
-
 	def before(self, label, pump_events=False):
-		"""Checks if the current trial time is before a given event.
-
-		.. warning:: Setting 'pump_events' to True will cause the input event queue to be cleared 
-		  every time the method is called. Please use :func:`~.pump` manually in loops where you
-		  will be processing user input.
+		"""Checks whether a given event has yet to occur in the trial.
 
 		Args:
-			label (str): The label of the event to check.
-			pump_events (bool, optional): If True, a :func:`~.ui_request` is performed when this
-				method is called. Defaults to False.
+			label (str): The name of the event to check.
+			pump_events (bool, optional): Deprecated, should always be False.
 
 		Returns:
-			True if the specified event has not yet occured within the trial, otherwise False.
+			True if the specified trial event has yet to occur, otherwise False.
 
 		"""
-		if type(label) is not str:
-			raise ValueError("Expected 'str' for argument label; got {0}.".format(type(label)))
-		if not self.registered(label):
-			raise NameError("Event '{0}' not registered with the EventManager.".format(label))
-		if pump_events:
-			ui_request()
-
-		return not self.__event_issued(label)
+		return not self.after(label, pump_events)
 
 
 	def after(self, label, pump_events=False):
-		"""Checks if the current trial time is after a given event.
-
-		.. warning:: Setting 'pump_events' to True will cause the input event queue to be cleared 
-		  every time the method is called. Please use :func:`~.pump` manually in loops where you
-		  will be processing user input.
+		"""Checks whether a given event has occurred in the trial.
 
 		Args:
-			label (str): The label of the event to check.
-			pump_events (bool, optional): If True, a :func:`~.ui_request` is performed when this
-				method is called. Defaults to False.
+			label (str): The name of the event to check.
+			pump_events (bool, optional): Deprecated, should always be False.
 
 		Returns:
-			True if the specified event has already occured within the trial, otherwise False.
+			True if the specified trial event has already occured, otherwise False.
 
 		"""
-		if type(label) is not str:
-			raise ValueError("Expected 'str' for argument label; got {0}.".format(type(label)))
-		if not self.registered(label):
-			raise NameError("Event '{0}' not registered with the EventManager.".format(label))
+		self._ensure_exists(label)
 		if pump_events:
 			ui_request()
 
-		return self.__event_issued(label)
+		# If event wasn't already issued, check if it should be issued now
+		if not self._issued[label]:
+			if self.events[label] < self.time_elapsed:
+				self._issued[label] = True
+
+		return self._issued[label]
 
 
-	def between(self, label_1, label_2):
-		"""Checks if the current trial time is between two given events.
+	def between(self, a, b):
+		"""Checks whether the current trial time is between two events.
 
 		Args:
-			label_1 (str): The label of the event to check if it is currently after.
-			label_2 (str): The label of the event to check if it is currently before.
+			a (str): The name of the first event.
+			b (str): The name of the second event.
 		
 		Returns:
-			True if the current trial time is between the specified events, otherwise False.
+			True if the trial time is between events a and b, otherwise False.
 
 		"""
-		return self.after(label_1) and self.before(label_2)
+		return self.after(a) and self.before(b)
 
 
-	def start_clock(self):
-		"""Starts the EventManager's trial clock. The onsets of all events registered with the
-		EventManager are relative to the point in time when this method is called.
+	def start(self):
+		"""Starts the EventManager's trial clock.
+		
+		The onsets of events added to the EventManager are relative to when this
+		method is called.
 
-		.. note:: This method is called automatically at the start of every trial. The only time
-		  you would ever need to call it manually is if you wanted to use the EventManager outside
-		  of an experiment's :meth:`~.KLExperiment.trial()` method.
-
-		Raises:
-			RuntimeError: If this method is called while the trial clock is already running.
+		If you are using the :class:`~klibs.KLExperiment.Experiment` object's
+		built-in ``self.evm``, this is called automatically at the start of
+		every trial.
 
 		"""
-		if self.start_time != None:
-			err = "The EventManger trial clock cannot be started if it is already running."
-			raise RuntimeError(err)
 		self.start_time = time()
+		for label in self.events.keys():
+			self._issued[label] = False
 
 
-	def stop_clock(self):
-		"""Stops the EventManager's trial clock and clears all registered events.
+	def reset(self):
+		"""Resets the EventManger, clearing all added events.
 
-		.. note:: This method is called automatically at the end of every trial. The only time
-		  you would ever need to call it manually is if you wanted to use the EventManager outside
-		  of an experiment's :meth:`~.KLExperiment.trial()` method.
+		If you are using the :class:`~klibs.KLExperiment.Experiment` object's
+		built-in ``self.evm``, this is called automatically at the end of
+		every trial.
 
 		"""
 		self.events = {}
+		self._issued = {}
 		self.start_time = None
+
+	
+	def start_clock(self):
+		# Alias for backwards compatibility, will be removed soon
+		self.start()
+
+
+	def stop_clock(self):
+		# Alias for backwards compatibility, will be removed soon
+		self.reset()
+
+
+	@property
+	def time_elapsed(self):
+		# Internal method for getting time since start() in milliseconds.
+		if not self.start_time:
+			return 0.0
+		return (time() - self.start_time) * 1000
+
 
 	@property
 	def trial_time(self):
-		"""float: Time in seconds since the start of the current trial. If called when the
-		EventManager clock is not running (e.g. outside of a trial), this will be equal to 0.
-		
-		"""
-		return 0.0 if self.start_time == None else (time() - self.start_time)
+		# Gets time since start() in seconds. Deprecated, not part of public API.
+		return self.time_elapsed / 1000.0
+
 
 	@property
 	def trial_time_ms(self):
-		"""float: Time in milliseconds since the start of the current trial. If called when the
-		EventManager clock is not running (e.g. outside of a trial), this will be equal to 0.
-		
-		"""
-		return self.trial_time * 1000
+		# Alias for backwards compatibility, not part of the public API.
+		return self.time_elapsed